+++ dtu/src/config.rs
@@ -1,276 +1,284 @@
-use std::fmt::Display;
-use std::path::Path;
 use std::{borrow::Cow, path::PathBuf};
-use toml::{Table, Value};
 
-use crate::utils::{path_must_str, read_file};
+use serde::Deserialize;
 
-#[derive(Debug)]
-pub enum Error {
-    InvalidType,
-    MissingKey,
+use crate::Context;
+
+#[derive(Deserialize, Clone)]
+pub struct AdbConfig {
+    serial: Option<String>,
+    executable: Option<String>,
 }
 
-impl Display for Error {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(
-            f,
-            "{}",
-            match self {
-                Self::InvalidType => "InvalidType",
-                Self::MissingKey => "MissingKey",
-            }
-        )
+impl Default for AdbConfig {
+    fn default() -> Self {
+        Self {
+            serial: None,
+            executable: None,
+        }
     }
 }
 
-pub type Result<T> = std::result::Result<T, Error>;
+impl AdbConfig {
+    pub fn get_serial(&self, ctx: &dyn Context) -> crate::Result<Cow<'_, str>> {
+        match self.serial.as_ref() {
+            Some(s) => Ok(Cow::Borrowed(s)),
+            None => ctx.get_env("ANDROID_SERIAL").map(Cow::Owned),
+        }
+    }
 
-pub struct ConfigMap<'c> {
-    path: &'c Path,
-    name: Option<Cow<'c, str>>,
-    table: &'c Table,
+    pub fn get_executable(&self, ctx: &dyn Context) -> crate::Result<Cow<'_, str>> {
+        match self.executable.as_ref() {
+            Some(s) => Ok(Cow::Borrowed(s)),
+            None => ctx.get_bin("adb").map(Cow::Owned),
+        }
+    }
 }
 
-#[derive(Clone)]
-pub struct Config {
-    path: PathBuf,
-    base: Table,
+const fn bool_true() -> bool {
+    true
 }
 
-impl Config {
-    pub fn parse(source: &Path) -> crate::Result<Self> {
-        let as_str = read_file(source)?;
+const fn bool_false() -> bool {
+    false
+}
 
-        let path = PathBuf::from(source);
+#[derive(Deserialize, Clone)]
+pub struct DumpConfig {
+    pub base: PathBuf,
+    #[serde(rename = "pull-is-link", default = "bool_false")]
+    pub pull_is_link: bool,
+}
 
-        let base: Table = match toml::from_str(&as_str) {
-            Ok(v) => v,
-            Err(e) => return Err(crate::Error::new_cfg(source, &e)),
+#[derive(Deserialize, Clone)]
+pub enum DeviceAccessConfig {
+    #[serde(rename = "adb")]
+    Adb(AdbConfig),
+    #[serde(rename = "dump")]
+    Dump(DumpConfig),
+}
+
+impl Default for DeviceAccessConfig {
+    fn default() -> Self {
+        Self::Adb(AdbConfig::default())
+    }
+}
+
+#[derive(Deserialize, Clone)]
+pub struct ProjectConfig {
+    #[serde(rename = "can-adb", default = "bool_true")]
+    pub can_adb: bool,
+
+    #[serde(rename = "device-access", default = "DeviceAccessConfig::default")]
+    pub device_access: DeviceAccessConfig,
+}
+
+impl Default for ProjectConfig {
+    fn default() -> Self {
+        Self {
+            can_adb: true,
+            device_access: DeviceAccessConfig::default(),
+        }
+    }
+}
+
+impl ProjectConfig {
+    pub fn is_adb(&self) -> bool {
+        matches!(self.device_access, DeviceAccessConfig::Adb(_))
+    }
+
+    pub fn get_adb_bin(&self, ctx: &dyn Context) -> crate::Result<Cow<'_, str>> {
+        if !self.can_adb {
+            return Err(crate::Error::AdbDisabled);
+        }
+
+        let DeviceAccessConfig::Adb(adb) = &self.device_access else {
+            return ctx.get_bin("adb").map(Cow::Owned);
         };
-        Ok(Self { base, path })
+
+        adb.get_executable(ctx)
+    }
+}
+
+#[derive(Clone, Deserialize)]
+pub struct LocalFileStoreConfig {
+    pub base: PathBuf,
+    #[serde(rename = "get-is-link", default = "bool_false")]
+    pub get_is_link: bool,
+}
+
+#[derive(Clone, Deserialize)]
+pub struct S3FileStoreConfig {
+    pub bucket: String,
+    profile: Option<String>,
+    #[serde(rename = "aws-bin")]
+    aws_bin: Option<String>,
+    #[serde(default = "bool_true")]
+    pub cache: bool,
+    #[serde(rename = "cache-is-link", default = "bool_false")]
+    pub cache_is_link: bool,
+}
+
+impl S3FileStoreConfig {
+    pub fn get_profile(&self, ctx: &dyn Context) -> Cow<'_, str> {
+        match &self.profile {
+            Some(v) => Cow::Borrowed(v),
+            None => ctx
+                .get_env("DTU_S3_PROFILE")
+                .map(Cow::Owned)
+                .unwrap_or_else(|_| Cow::Borrowed("dtu")),
+        }
     }
 
-    pub fn invalid_error(&self, msg: String) -> crate::Error {
-        crate::Error::InvalidConfig(path_must_str(&self.path).into(), msg)
-    }
-
-    pub fn get_map(&self) -> ConfigMap<'_> {
-        ConfigMap {
-            name: None,
-            path: &self.path,
-            table: &self.base,
+    pub fn get_aws_bin(&self, ctx: &dyn Context) -> crate::Result<Cow<'_, str>> {
+        match &self.aws_bin {
+            Some(v) => Ok(Cow::Borrowed(v)),
+            None => ctx.get_bin("aws").map(Cow::Owned),
         }
     }
 }
 
-pub fn parse_config<R, F>(file: &Path, f: F) -> crate::Result<R>
-where
-    F: FnOnce(&ConfigMap) -> crate::Result<R>,
-{
-    let as_str = read_file(file)?;
+impl LocalFileStoreConfig {
+    pub fn new(base: PathBuf, get_is_link: bool) -> Self {
+        Self { base, get_is_link }
+    }
 
-    let table: Table = match toml::from_str(&as_str) {
-        Ok(v) => v,
-        Err(e) => return Err(crate::Error::new_cfg(file, &e)),
-    };
-
-    let base = ConfigMap {
-        name: None,
-        path: file,
-        table: &table,
-    };
-
-    let res = f(&base);
-
-    res
+    fn get_default(ctx: &dyn Context) -> crate::Result<Self> {
+        let base = ctx.get_user_local_dir()?.join("filestore");
+        let get_is_link = false;
+        Ok(Self::new(base, get_is_link))
+    }
 }
 
-impl<'c> ConfigMap<'c> {
-    fn get_full_path<'a>(&'a self) -> Option<&'a str> {
-        self.name.as_ref().map(|it| it.as_ref())
-    }
+#[derive(Clone, Deserialize)]
+pub enum FileStoreConfig {
+    #[serde(rename = "local")]
+    Local(LocalFileStoreConfig),
+    #[serde(rename = "s3")]
+    S3(S3FileStoreConfig),
+}
 
-    fn key_path<'a>(&self, key: &'a str) -> Cow<'a, str> {
-        match self.get_full_path() {
-            None => Cow::Borrowed(key),
-            Some(v) => Cow::Owned(format!("{}.{}", v, key)),
-        }
+impl FileStoreConfig {
+    fn get_default(ctx: &dyn Context) -> crate::Result<Self> {
+        Ok(Self::Local(LocalFileStoreConfig::get_default(ctx)?))
     }
+}
 
-    /// Helper to create a crate::Error for a missing key
-    pub fn missing_key(&self, key: &str) -> crate::Error {
-        let path = self.key_path(key);
-        crate::Error::InvalidConfig(
-            path_must_str(self.path).into(),
-            format!("missing key: {}", path),
-        )
-    }
+#[derive(Clone, Deserialize)]
+pub struct GlobalConfig {
+    pub filestore: FileStoreConfig,
+}
 
-    /// Helper to create a crate::Error for an invalid key
-    pub fn invalid_key(&self, key: &str, expected: &str) -> crate::Error {
-        let path = self.key_path(key);
-        crate::Error::InvalidConfig(
-            path_must_str(self.path).into(),
-            format!(
-                "invalid value for key: {} (expected type: {})",
-                path, expected
-            ),
-        )
-    }
-
-    pub fn has(&self, key: &str) -> bool {
-        self.table.contains_key(key)
-    }
-
-    fn get(&self, key: &str) -> Option<&'c Value> {
-        self.table.get(key)
-    }
-
-    pub fn maybe_get_int(&self, key: &str) -> Result<Option<i64>> {
-        match self.get(key) {
-            Some(v) => match v.as_integer() {
-                Some(v) => Ok(Some(v)),
-                None => Err(Error::InvalidType),
-            },
-            None => Ok(None),
-        }
-    }
-
-    pub fn maybe_get_int_typecheck(&self, key: &str) -> crate::Result<Option<i64>> {
-        self.maybe_get_int(key)
-            .map_err(|_| self.invalid_key(key, "int"))
-    }
-
-    pub fn get_str(&self, key: &str) -> Result<&'c str> {
-        self.maybe_get_str(key)?.ok_or(Error::MissingKey)
-    }
-
-    pub fn maybe_get_str(&self, key: &str) -> Result<Option<&'c str>> {
-        match self.get(key) {
-            Some(v) => match v.as_str() {
-                Some(v) => Ok(Some(v)),
-                None => Err(Error::InvalidType),
-            },
-            None => Ok(None),
-        }
-    }
-
-    pub fn maybe_get_str_typecheck(&self, key: &str) -> crate::Result<Option<&'c str>> {
-        self.maybe_get_str(key)
-            .map_err(|_| self.invalid_key(key, "string"))
-    }
-
-    pub fn must_get_str(&self, key: &str) -> crate::Result<&'c str> {
-        match self.get_str(key) {
-            Err(Error::InvalidType) => Err(self.invalid_key(key, "string")),
-            Err(Error::MissingKey) => Err(self.missing_key(key)),
-            Ok(v) => Ok(v),
-        }
-    }
-
-    pub fn get_bool(&self, key: &str) -> Result<bool> {
-        self.get(key)
-            .ok_or(Error::MissingKey)?
-            .as_bool()
-            .ok_or(Error::InvalidType)
-    }
-
-    pub fn get_int(&self, key: &str) -> Result<i64> {
-        self.get(key)
-            .ok_or(Error::MissingKey)?
-            .as_integer()
-            .ok_or(Error::InvalidType)
-    }
-
-    pub fn get_bool_or(&self, key: &str, default: bool) -> bool {
-        self.table
-            .get(key)
-            .map(|it| it.as_bool().unwrap_or(default))
-            .unwrap_or(default)
-    }
-
-    pub fn maybe_get_map(&'c self, key: &'c str) -> Result<Option<ConfigMap<'c>>> {
-        match self.get(key) {
-            Some(v) => match v.as_table() {
-                Some(table) => {
-                    let name = match &self.get_full_path() {
-                        Some(parents) => Cow::Owned(format!("{parents}.{key}")),
-                        None => Cow::Borrowed(key),
-                    };
-                    Ok(Some(Self {
-                        name: Some(name),
-                        path: self.path,
-                        table,
-                    }))
-                }
-                None => Err(Error::InvalidType),
-            },
-            None => Ok(None),
-        }
-    }
-
-    pub fn get_map(&'c self, key: &'c str) -> Result<ConfigMap<'c>> {
-        self.maybe_get_map(key)?.ok_or(Error::MissingKey)
-    }
-
-    pub fn maybe_get_map_typecheck(&'c self, key: &'c str) -> crate::Result<Option<ConfigMap<'c>>> {
-        self.maybe_get_map(key)
-            .map_err(|_| self.invalid_key(key, "table"))
-    }
-
-    pub fn must_get_map(&'c self, key: &'c str) -> crate::Result<ConfigMap<'c>> {
-        match self.get_map(key) {
-            Err(Error::InvalidType) => Err(self.invalid_key(key, "table")),
-            Err(Error::MissingKey) => Err(self.missing_key(key)),
-            Ok(v) => Ok(v),
-        }
+impl GlobalConfig {
+    pub fn get_default(ctx: &dyn Context) -> crate::Result<Self> {
+        Ok(Self {
+            filestore: FileStoreConfig::get_default(ctx)?,
+        })
     }
 }
 
 #[cfg(test)]
 mod test {
 
-    use crate::testing::{global_tmp_context, TestContext};
-    use std::fs;
-    use std::path::PathBuf;
+    use std::sync::atomic::{AtomicBool, Ordering};
+
+    use crate::testing::{mock_context, MockContext};
 
     use super::*;
     use rstest::*;
 
-    #[fixture]
-    #[once]
-    fn cfg_file(global_tmp_context: &TestContext) -> PathBuf {
-        let path = global_tmp_context.get_temp_dir();
-        fs::create_dir_all(&path).expect("create tmp dir");
-        let file = path.join("config.toml");
-
-        let content = r#"
-base = 12
-
-[foo]
-bar = "baz"
-quux = { neato = true }
+    #[rstest]
+    fn test_project_config_adb(mut mock_context: MockContext) {
+        let raw_config = r#"
+[device-access.adb]
+serial = "DEVICE_SERIAL"
+executable = "/path/to/adb"
 "#;
 
-        fs::write(&file, content).expect("failed to write test config");
-        file
+        let config: ProjectConfig = toml::from_str(raw_config).expect("parse config");
+        assert!(config.can_adb, "should have adb enabled by default");
+
+        let DeviceAccessConfig::Adb(adb) = config.device_access else {
+            panic!("should be adb device access");
+        };
+        assert_eq!(adb.get_serial(&mock_context).unwrap(), "DEVICE_SERIAL");
+        assert_eq!(adb.get_executable(&mock_context).unwrap(), "/path/to/adb");
+
+        let raw_config = "";
+        let config: ProjectConfig = toml::from_str(raw_config).expect("parse config");
+        assert!(config.can_adb, "should have adb enabled by default");
+
+        let DeviceAccessConfig::Adb(adb) = config.device_access else {
+            panic!("should default to adb device access");
+        };
+
+        let first_ser = AtomicBool::new(true);
+        let first_bin = AtomicBool::new(true);
+
+        mock_context.expect_maybe_get_env().returning(move |e| {
+            if e == "ANDROID_SERIAL" {
+                if first_ser.load(Ordering::Relaxed) {
+                    first_ser.store(false, Ordering::Relaxed);
+                    None
+                } else {
+                    Some(String::from("suchserial"))
+                }
+            } else {
+                None
+            }
+        });
+
+        mock_context.expect_maybe_get_bin().returning(move |b| {
+            if b == "adb" {
+                if first_bin.load(Ordering::Relaxed) {
+                    first_bin.store(false, Ordering::Relaxed);
+                    None
+                } else {
+                    Some(String::from("/usr/bin/adb"))
+                }
+            } else {
+                None
+            }
+        });
+
+        assert!(adb.get_serial(&mock_context).is_err());
+        assert!(adb.get_executable(&mock_context).is_err());
+
+        assert_eq!(adb.get_serial(&mock_context).unwrap(), "suchserial");
+        assert_eq!(adb.get_executable(&mock_context).unwrap(), "/usr/bin/adb");
     }
 
     #[rstest]
-    fn test_config(cfg_file: &PathBuf) {
-        let f = |cfg: &ConfigMap| -> crate::Result<()> {
-            assert_eq!(cfg.get_int("base").expect("getting base"), 12);
-            let foo = cfg.get_map("foo").expect("getting foo");
-            assert_eq!(foo.get_str("bar").expect("getting bar"), "baz");
-            assert!(foo.get_str("ohno").is_err());
-            let quux = foo.get_map("quux").expect("getting foo.quux");
-            assert_eq!(quux.get_bool("neato").expect("getting neato"), true);
-            assert_eq!(quux.get_bool_or("ohno", false), false);
-            assert_eq!(quux.get_bool_or("ohno", true), true);
-            Ok(())
+    fn test_project_config_dump() {
+        let raw_config = r#"
+can-adb = false
+[device-access.dump]
+base = "/path/to/fs/dump/root"
+pull-is-link = false
+"#;
+
+        let config: ProjectConfig = toml::from_str(raw_config).expect("parse config");
+        assert!(!config.can_adb, "adb should be disabled");
+
+        let DeviceAccessConfig::Dump(dump) = config.device_access else {
+            panic!("should have been dump device access");
+        };
+        assert_eq!(dump.base, PathBuf::from("/path/to/fs/dump/root"));
+        assert!(!dump.pull_is_link);
+    }
+
+    #[rstest]
+    fn test_global_config_s3(mock_context: MockContext) {
+        let raw = r#"[filestore.s3]
+bucket = "neato"
+aws-bin = "aws"
+"#;
+        let config: GlobalConfig = toml::from_str(&raw).expect("parse config");
+        let FileStoreConfig::S3(s3) = config.filestore else {
+            panic!("should have been a file store config");
         };
 
-        parse_config(cfg_file, f).unwrap();
+        assert_eq!(s3.get_aws_bin(&mock_context).unwrap(), "aws");
+        assert_eq!(s3.bucket, "neato");
     }
 }