+++ CHANGELOG.md
@@ -15,6 +15,7 @@ This new version contains significant breaking changes to the graph database and
 - **BREAKING** Made `FileStore` require `Send` and `Sync`
 - **BREAKING** `meta set-progress` now sets to true and `clear-progress` added to set to false
 - **BREAKING** Completely removed the `DeviceDatabase` trait. This database will always be `sqlite` and it just added complexity and actually make things less efficient. `DeviceSqliteDatabase` was renamed to `DeviceDatabase`
+- **BREAKING** Reworked configuration
 - Exposed the `SqliteConnection` on the `DeviceDatabase` and `SqliteGraphDatabase` for direct access
 - Added some methods to `DeviceDatabase`
 - Create Python bindings for the databases and the application server